requirements.txt
confluent-kafka
tensorflow
tensorflow-io
python-dotenv
pandas